add grenade

+++ Assets/script/PlayerGameManager.cs
@@ -209,7 +209,7 @@ public class PlayerGameManager : MonoBehaviour {
 		}
 	}
 		
-	public void takeDamage(int damage){
+	public void takeDamage(float damage){
 		health -= damage;
 		if (health < 0f) { health = 0;}
